+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v1.3.4
   should be thrown instead.
 * Fix deadlock between node split and checkpoint.
 * Fix "Already in NodeMap" exception when loading fragmented nodes.
+* Fix for a rare assertion error when deleting the root node of a tree.
 * Added a Cursor lock method, for manual lock control.
 * Added some default View and Cursor method implementations.